Prepares results data per asset type for bubble chart

This function aggregates and prepares one of PACTA analysis result file "Bonds/Equity_results_portfolio.rda" from "40_Results" folder to form an input that can be used for plotting meta-report bubble chart showing a planned technology build-out as percentage of build-out required by a scenario against the current technology share.

prepare_for_metareport_bubble_chart(
  data_asset_type,
  asset_type,
  start_year,
  technologies_filter,
  scenario_filter,
  scenario_geography_filter
)

Arguments

data_asset_type
Dataframe in the shape of "Equity/Bonds_results_portfolio.rda" dataset from PACTA analysis output in "40_Results" folder (dataframe).

asset_type
Asset type of the data (character string).

start_year
Year to be used as a start year in difference calculations (integer).

technologies_filter
Technologies to be used for filtering (character string or a vector of character strings).

scenario_filter
Scenario to be used in the calculations (character string).

scenario_geography_filter
Scenario geography to be used in the calculations (character string)

Value

Dataframe with columns investor_name, portfolio_name, value_x, value_y.

Examples

# TODO