Time series of emission factors

An example data set containing emission factors for different metric types: projected, benchmark and scenarios and different sectors.

sda_target

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 208 rows and 4 columns.

Source

https://github.com/2DegreesInvesting/r2dii.plot.static/issues/55.

Details

Column definitions:

sector: Industry sector (as defined in PACTA).

year: Year for which emissions are reported.

emission_factor_metric: metric type: 'projected' for the test portfolio 'corporate_economy' for the benchmark, 'adjusted_scenario_demo' for a scenario.

emission_factor_value: emissions value, a double, units differ depending on the sector.

Examples

sda_target

#> # A tibble: 208 x 4
#>    sector      year emission_factor_metric emission_factor_value
#>    <chr>      <dbl> <chr>                                  <dbl>
#>  1 automotive  2002 projected                              0.228
#>  2 automotive  2003 projected                              0.226
#>  3 automotive  2004 projected                              0.224
#>  4 automotive  2005 projected                              0.222
#>  5 automotive  2006 projected                              0.220
#>  6 automotive  2007 projected                              0.218
#>  7 automotive  2008 projected                              0.216
#>  8 automotive  2009 projected                              0.214
#>  9 automotive  2010 projected                              0.212
#> 10 automotive  2011 projected                              0.210
#> # … with 198 more rows